Move EthTx's high-level formatting logic into adapters pkg
Also, better documentation
