+++ core/utils/ethabi.go
@@ -3,7 +3,6 @@ package utils
 import (
 	"bytes"
 	"encoding/binary"
-	"encoding/hex"
 	"fmt"
 	"math/big"
 	"strconv"
@@ -12,35 +11,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/tidwall/gjson"
-)
-
-const (
-	// FormatRawHexWithFuncSelectorAndDataPrefix is the default format behavior.
-	// Its output is prefixed with the ethTx function selector, followed by the
-	// data prefix
-	FormatRawHexWithFuncSelectorAndDataPrefix = ""
-	// FormatBytes encodes the output as bytes. I.e., the string input will be
-	// cast to bytes, and passed to the on-chain contract method as a solidity
-	// bytes array argument. (No conversion from hex is done; the string input
-	// must be the raw bytes!)
-	FormatBytes = "bytes"
-	// FormatPreformatted encodes the output, assumed to be hex, as bytes and
-	// passes them as arguments. Caller is responsible for all formatting for the
-	// EVM. Input must be 0x-prefixed
-	FormatPreformattedHexArguments = "preformattedHexArguments"
-	// FormatRawHex does no formatting at all. Caller is responsible for
-	// formatting the function selector and offset, in addition to any arguments
-	// to be passed with the transaction. Input must be 0x-prefixed
-	//
-	// Note that this option isn't adressed in EVMTranscodeJSONWithFormat, because
-	// eth_tx.go's getTxData short-circuits, when it encounters this.
-	FormatRawHex = "rawHex"
-	// FormatUint256 encodes the output as bytes containing a uint256
-	FormatUint256 = "uint256"
-	// FormatInt256 encodes the output as bytes containing an int256
-	FormatInt256 = "int256"
-	// FormatBool encodes the output as bytes containing a bool
-	FormatBool = "bool"
 )
 
 // ConcatBytes appends a bunch of byte arrays into a single byte array
@@ -196,44 +166,6 @@ func EVMTranscodeInt256(value gjson.Result) ([]byte, error) {
 	}
 
 	return EVMWordSignedBigInt(output)
-}
-
-// EVMTranscodeJSONWithFormat given a JSON input and a format specifier, encode the
-// value for use by the EVM
-func EVMTranscodeJSONWithFormat(value gjson.Result, format string) ([]byte, error) {
-	switch format {
-	case FormatBytes:
-		return EVMTranscodeBytes(value)
-	case FormatPreformattedHexArguments:
-		if !HasHexPrefix(value.Str) {
-			return nil, fmt.Errorf("%s input must be 0x-prefixed, got %s",
-				FormatPreformattedHexArguments, value.Str)
-		}
-		return hex.DecodeString(RemoveHexPrefix(value.Str))
-	case FormatUint256:
-		data, err := EVMTranscodeUint256(value)
-		if err != nil {
-			return []byte{}, err
-		}
-		return EVMEncodeBytes(data), nil
-
-	case FormatInt256:
-		data, err := EVMTranscodeInt256(value)
-		if err != nil {
-			return []byte{}, err
-		}
-		return EVMEncodeBytes(data), nil
-
-	case FormatBool:
-		data, err := EVMTranscodeBool(value)
-		if err != nil {
-			return []byte{}, err
-		}
-		return EVMEncodeBytes(data), nil
-
-	default:
-		return []byte{}, fmt.Errorf("unsupported format: %s", format)
-	}
 }
 
 // EVMWordUint64 returns a uint64 as an EVM word byte array.
